finished

+++ ToDoList/src/components/ToDoWrapper.jsx
@@ -2,10 +2,10 @@ import React, { useState, useEffect } from 'react';
 import { ToDoForm } from './ToDoForm';
 import { v4 as uuidv4 } from 'uuid';
 import { ToDo } from './ToDo';
-import { EditDoForm } from './EditToDoForm';
+import { EditToDoForm } from './EditToDoForm';
 import { doSignOut } from './firebase/auth';
 import { useNavigate } from 'react-router-dom';
-import { collection, addDoc, updateDoc, doc, deleteDoc , onSnapshot} from "firebase/firestore";
+import { collection, addDoc, updateDoc, doc, deleteDoc ,getDoc, onSnapshot} from "firebase/firestore";
 import { db } from "./firebase/firebase"; 
 uuidv4();
 
@@ -22,6 +22,7 @@ export const ToDoWrapper = () => {
     }
     return [];
   });
+  const [isEditing, setIsEditing] = useState(false)
 
   const navigate = useNavigate();
   useEffect(() => {
@@ -40,23 +41,22 @@ export const ToDoWrapper = () => {
     localStorage.setItem('todos', JSON.stringify(todoArray));
   }, [todoArray]);
 
-  const addTodo = async(todoName) => {
+  const addTodo = async (todoName) => {
     const newTodo = {
-      id: uuidv4(),
       task: todoName,
       completed: false,
       isEditing: false,
     };
+    console.log(todoArray)
     try {
+     
       const docRef = await addDoc(collection(db, "todos"), newTodo);
       console.log("Document written with ID: ", docRef.id);
-  
-      // Update state with the new todo
-      setTodos((prevTodos) => [...prevTodos, newTodo]);
+      
+      
     } catch (e) {
       console.error("Error adding document: ", e);
     }
-    setTodos((prevTodos) => [...prevTodos, newTodo]);
   };
 
   const toggleComplete = async(id) => {
@@ -65,7 +65,6 @@ export const ToDoWrapper = () => {
       const updatedTodo = todoArray.find((todo) => todo.id === id);
       await updateDoc(todoDoc, { completed: !updatedTodo.completed });
   
-      // Update state locally
       setTodos((prevTodos) =>
         prevTodos.map((todo) =>
           todo.id === id ? { ...todo, completed: !todo.completed } : todo
@@ -86,34 +85,58 @@ export const ToDoWrapper = () => {
   const deleteTodo = async(id) => {
     try {
       const todoDoc = doc(db, "todos", id);
-      await deleteDoc(todoDoc);
-  
-      // Update state locally
-      setTodos((prevTodos) => prevTodos.filter((todo) => todo.id !== id));
+      
+      const docSnap = await getDoc(todoDoc);
+      if (docSnap.exists()) {
+        await deleteDoc(todoDoc);
+        console.log("Todo deleted from Firestore");
+        setTodos((prevTodos) => prevTodos.filter((todo) => todo.id !== id));
+      } else {
+        console.log("Todo not found in Firestore");
+      }
     } catch (e) {
       console.error("Error deleting document: ", e);
     }
-    setTodos((prevTodos) => prevTodos.filter((todoObject) => todoObject.id !== id));
+  };
+  
+
+  const editTodo = async(id) => {
+    try {
+      setIsEditing(true)
+      const todoToEdit = todoArray.find((todoObject) => todoObject.id === id);
+      if (!todoToEdit) return;
+  
+      const todoDoc = doc(db, "todos", id);
+      await updateDoc(todoDoc, { isEditing: !todoToEdit.isEditing });
+  
+      setTodos((prevTodos) =>
+        prevTodos.map((todoObject) =>
+          todoObject.id === id
+            ? { ...todoObject, isEditing: !todoObject.isEditing }
+            : todoObject
+        )
+      );
+    } catch (e) {
+      console.error("Error updating document:", e);
+    }
   };
 
-  const editTodo = (id) => {
-    setTodos((prevTodos) =>
-      prevTodos.map((todoObject) =>
-        todoObject.id === id
-          ? { ...todoObject, isEditing: !todoObject.isEditing }
-          : todoObject
-      )
-    );
-  };
-
-  const editTask = (todoName, id) => {
-    setTodos((prevTodos) =>
-      prevTodos.map((todoObject) =>
-        todoObject.id === id
-          ? { ...todoObject, task: todoName, isEditing: !todoObject.isEditing }
-          : todoObject
-      )
-    );
+  const editTask = async(todoName, id) => {
+    try {
+      const todoDoc = doc(db, "todos", id);
+      await updateDoc(todoDoc, { task: todoName, isEditing: false }); 
+  
+      setTodos((prevTodos) =>
+        prevTodos.map((todoObject) =>
+          todoObject.id === id
+            ? { ...todoObject, task: todoName, isEditing: false }
+            : todoObject
+        )
+      );
+      setIsEditing(false)
+    } catch (e) {
+      console.error("Error updating document:", e);
+    }
   };
 
   return (
@@ -130,8 +153,8 @@ export const ToDoWrapper = () => {
       <h1>Get Things Done!</h1>
       <ToDoForm addTodo={addTodo} />
       {todoArray.map((todoObject) =>
-        todoObject.isEditing ? (
-          <EditDoForm editTodo={editTask} task={todoObject} key={todoObject.id} />
+        isEditing ? (
+          <EditToDoForm editTodo={editTask} task={todoObject} key={todoObject.id} />
         ) : (
           <ToDo
             task={todoObject}
